feat: refactoring project to use proper inversion of control
