+++ src/modules/sync/repository/tableDefinitions.ts
