+++ src/modules/sync/sync/tables/abstractFileSyncTable.ts
@@ -1,5 +1,5 @@
 import { App, TFile } from "obsidian";
-import { SqlSealDatabase } from "../../modules/database/database";
+import { SqlSealDatabase } from "../../../database/database";
 
 export abstract class AFileSyncTable {
     constructor(
