+++ types-package/src/index.ts
@@ -1,4 +1,4 @@
-import type { SQLSealRegisterApi, SQLSealApi  } from '../../src/pluginApi/sqlSealApi'
+import type { SQLSealRegisterApi, SQLSealApi  } from '../../src/modules/api/pluginApi/sqlSealApi'
 import { RendererConfig } from '../../src/modules/editor/renderer/rendererRegistry'
